sendfile/f2s_sndtimeo: add type cast
Previously difference of two unsigned integers is always nonnegative.

+++ sockapi-ts/sendfile/f2s_sndtimeo.c
@@ -117,7 +117,7 @@ main(int argc, char *argv[])
                             file_length - offset, FALSE);
 
         RING("sendfile() blocks within %d microseconds", pco_iut->duration);
-        if (abs(sndtimeo - pco_iut->duration / 1000) > 1)
+        if (abs(sndtimeo - (int)(pco_iut->duration / 1000)) > 1)
             WARN("Duration of sendfile() processing is not within "
                  "requested interval in msec: requested=%d, actually=%d",
                  sndtimeo, pco_iut->duration / 1000);
